backend/app/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
gunicorn
requests
Pillow
django-filter
django-extensions
celery
redis
openai
pandas
numpy
python-dateutil
pytz
